changed some styles
