+++ src/App.tsx
@@ -1,15 +1,16 @@
-// import { StartPage } from './components/StartPage/StartPage';
-import { MainQuizPage } from './components/MainQuizPage/MainQuizPage';
+import { StartPage } from './components/pages/StartPage/StartPage';
+import { MainQuizPage } from './components/pages/MainQuizPage/MainQuizPage';
 
 import './App.css';
 
 function App() {
 
-  return (
-    <>
-      <MainQuizPage />
-    </>
-  )
+    return (
+        <>
+            <StartPage />
+            <MainQuizPage />
+        </>
+    )
 }
 
 export { App };
