--- src/components/MainQuizPage/EndQuizButton/EndQuizButton.tsx
@@ -1,7 +1,0 @@
-const EndQuizButton = () => {
-    return (
-        <button type="button" className="btn btn-primary">Finish test</button>
-    );
-};
-
-export { EndQuizButton };
