--- src/components/StartPage/ButtonsContainer/ButtonsContainer.tsx
@@ -1,10 +1,0 @@
-const ButtonsContainer = () => {
-    return (
-        <div className="flex-row buttons-container">
-            <button type="button">Start quiz</button>
-            <button type="button">See my stats</button>
-        </div>
-    );
-};
-
-export { ButtonsContainer };
