+++ src/components/pages/MainQuizPage/EndQuizButton/EndQuizButton.tsx
@@ -1,0 +1,9 @@
+import { StyledButton } from "../../../spreadedComponents/StyledButton/StyledButton";
+
+const EndQuizButton = () => {
+    return (
+        <StyledButton text="Finish test" />
+    );
+};
+
+export { EndQuizButton };
