+++ src/components/pages/MainQuizPage/MainQuizPage.tsx
