+++ src/components/pages/MainQuizPage/ProgressBarContainer/ProgressBarContainer.tsx
