+++ src/components/pages/MainQuizPage/ProgressBarContainer/interface/progressBarContainer.interface.ts
