+++ src/components/pages/MainQuizPage/QuestionContainer/ButtonsContainer/AnswerButton/interface/answerButton.interface.ts
