+++ src/components/pages/MainQuizPage/QuestionContainer/ButtonsContainer/GroupButtonsContainer/interface/groupButtonsContainer.ts
