+++ src/components/pages/MainQuizPage/QuestionContainer/ButtonsContainer/interfaces/questionType.interface.ts
