+++ src/components/pages/MainQuizPage/QuestionContainer/interface/QuestionContainer.interface.ts
