+++ src/components/pages/MainQuizPage/QuestionContainer/interface/question.interface.ts
