+++ src/components/pages/MainQuizPage/TimerContainer/TimerContainer.tsx
