+++ src/components/pages/MainQuizPage/TimerContainer/interface/timerContainer.interface.ts
