+++ src/components/pages/StartPage/NumberInputContainer/NumberInput/NumberInput.tsx
