+++ src/components/pages/StartPage/NumberInputContainer/NumberInput/interfaces/NumberInput.interface.ts
