+++ src/components/pages/StartPage/NumberInputContainer/NumberInputContainer.tsx
