+++ src/components/pages/StartPage/SelectsContainer/SelectInput/SelectInput.tsx
