+++ src/components/pages/StartPage/SelectsContainer/SelectInput/interfaces/SelectInput.interface.ts
