+++ src/components/pages/StartPage/SelectsContainer/options/option.interface.ts
