+++ src/components/pages/StartPage/StartPage.tsx
