+++ src/components/spreadedComponents/StyledButton/StyledButton.tsx
@@ -1,0 +1,9 @@
+import { styledButtonProps } from "./interface/styledButton.interface"
+
+const StyledButton = ({ text }: styledButtonProps) => {
+    return (
+        <button className="styled-button" type="button">{text}</button>
+    )
+}
+
+export { StyledButton };
