+++ src/components/spreadedComponents/StyledButton/interface/styledButton.interface.ts
@@ -1,0 +1,3 @@
+export interface styledButtonProps {
+    text: string
+}
