+++ src/index.css
@@ -55,26 +55,8 @@ body {
   justify-content: center;
 }
 
-.buttons-container button {
-  margin-top: 5px;
-  margin-right: 20px;
-  padding: 5px 15px 5px 15px;
-  display: block;
-  font-size: 20px;
-  color: #fff;
-  border: none;
-  border-radius: 10px;
-  background-image: linear-gradient(to right, #aa076b, #61045f);
-}
-
 label {
   color: #2f4f4f;
-}
-
-.buttons-container button:hover {
-  background-image: linear-gradient(to right, #61045f, #aa076b);
-  font-size: 18px;
-  cursor: pointer;
 }
 
 .select-input {
@@ -139,7 +121,7 @@ label {
   justify-content: space-between;
 }
 
-.main-quiz-page-footer button {
+.styled-button {
   margin-top: 5px;
   margin-right: 20px;
   padding: 5px 15px 5px 15px;
@@ -151,8 +133,7 @@ label {
   background-color: #aa076b;
 }
 
-.main-quiz-page-footer button:hover {
+.styled-button:hover {
   background-color: #61045f;
   cursor: pointer;
 }
-
